+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Solid integer".
 - Tests for "Bit count to limit".
 - Tests for "Porcupine fever".
+- Tests for "Bijective numeration".
 
 ## [1.15.0] - 2022-11-30
 ### Added